Fix timing, extract static method
