+++ HUMR/Scenes/SampleScene_UdonProgramSources/UdonUtilities.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7585df8e039e0c145afe5a27ed1964df
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
