+++ HUMR/Scripts/Csharp/OutputLogLoader.cs
@@ -1,4 +1,3 @@
-
 /*******
  * OutputLogLoader.cs
  *
