+++ HUMR/Scripts/Editor/HUMRImportFBXSettings.cs
@@ -1,4 +1,3 @@
-
 /*******
  * HUMRImportFBXSettings.cs
  * 
@@ -11,6 +10,7 @@
  * 上記の設定に関して不具合等が発生したらこのスクリプトを改造するか削除して対応のこと
  * 
  * *****/
+
 #if UNITY_EDITOR
 using UnityEngine;
 using UnityEditor;
