+++ HUMR/Scripts/Editor/OutputLogLoaderEditor.cs
@@ -1,4 +1,3 @@
-
 /*******
  * OutputLogLoaderEditor.cs
  * 
@@ -10,6 +9,7 @@
  * "LoadLogToExportAnim"のボタンを押すとOutputLog内の処理が実行される
  * 
  * *****/
+
 #if UNITY_EDITOR
 using UnityEngine;
 using UnityEditor;
