+++ HUMR/Scripts/UdonSharp/InteractRecorder.cs
@@ -1,4 +1,3 @@
-
 /*******
  * InteractRecorder.cs
  *
