+++ HUMR/Scripts/UdonSharp/Recorder.cs
@@ -1,4 +1,3 @@
-
 /*******
  * Recorder.cs
  * 
@@ -8,12 +7,11 @@
  * 
  * *****/
 
-using System.Globalization;
-using System.Text;
 using UdonSharp;
 using UnityEngine;
 using VRC.SDKBase;
 using VRC.Udon;
+
 
 namespace HUMR
 {
@@ -27,14 +25,15 @@ namespace HUMR
 
         private VRCPlayerApi[] _players;
         private VRCPlayerApi _localPlayer;
-        private float _startTime;
+        private float _recordTime;
         private float _recordInterval;
         private float _nextRecordTime;
 
         private void Start()
         {
             _localPlayer = Networking.LocalPlayer;
-            _startTime = Time.time;
+            _recordTime = 0f;
+            _nextRecordTime = _recordTime;
             _recordInterval = 1f / recordFramerate;
             _players = new VRCPlayerApi[10];
             RefreshPlayerList();
@@ -42,69 +41,22 @@ namespace HUMR
 
         private void Update()
         {
-            var time = Time.time - _startTime;
-            if (time < _nextRecordTime) return;
-
-            _nextRecordTime = Time.time + _recordInterval;
+            _recordTime += Time.deltaTime;
+            if (_recordTime < _nextRecordTime) return;
+            _nextRecordTime = _recordTime + _recordInterval;
 
             if (recordAllPlayers)
             {
                 foreach (var player in _players)
                 {
                     if (!Utilities.IsValid(player)) continue;
-                    LogPlayer(player, time);
+                    UdonUtilities.LogPlayerBoneRotations(player, _recordTime);
                 }
             }
             else
             {
-                LogPlayer(_localPlayer, time);
+                UdonUtilities.LogPlayerBoneRotations(_localPlayer, _recordTime);
             }
-            
-            var playerCount = VRCPlayerApi.GetPlayerCount();
-            for (var i = 0; i < playerCount; i++)
-            {
-                if (!recordAllPlayers && i > 0) break;
-                
-                var player = _players[i];
-                if (!Utilities.IsValid(player)) continue;
-
-                LogPlayer(player, time);
-            }
-        }
-
-        private static void LogPlayer(VRCPlayerApi player, float time)
-        {
-            var builder = new StringBuilder(4096);
-
-            builder.Append("HUMR:");
-            builder.Append(player.displayName);
-            builder.Append(';');
-            builder.Append(time.ToString(CultureInfo.InvariantCulture));
-
-            var hipsPosition = player.GetBonePosition(HumanBodyBones.Hips);
-
-            builder.Append(';');
-            builder.Append(hipsPosition.x.ToString("F7", CultureInfo.InvariantCulture));
-            builder.Append(',');
-            builder.Append(hipsPosition.y.ToString("F7", CultureInfo.InvariantCulture));
-            builder.Append(',');
-            builder.Append(hipsPosition.z.ToString("F7", CultureInfo.InvariantCulture));
-
-            for (var i = 0; i < HumanTrait.BoneName.Length; i++)
-            {
-                var rotation = player.GetBoneRotation((HumanBodyBones)i);
-
-                builder.Append(';');
-                builder.Append(rotation.x.ToString("F7", CultureInfo.InvariantCulture));
-                builder.Append(',');
-                builder.Append(rotation.y.ToString("F7", CultureInfo.InvariantCulture));
-                builder.Append(',');
-                builder.Append(rotation.z.ToString("F7", CultureInfo.InvariantCulture));
-                builder.Append(',');
-                builder.Append(rotation.w.ToString("F7", CultureInfo.InvariantCulture));
-            }
-
-            Debug.Log(builder.ToString());
         }
 
         public override void OnPlayerJoined(VRCPlayerApi player)
